Break ServiceInstance into Instance and Container
In the mesh use-case, this split allows seperation of concerns.

The container model is maintained by executor and contains information
about the container and its state including what is deployed within it.

The instance model is maintained by the strong-pm instance running
within the container. All the information on this model reflects the
current live state of the PM.

Both models are meant to be read-only on the REST API.