server-service: deploy() needs raw request
Using the http context argument, all the data has been read from the req
object, making it unuseable for deploy. Using an explicit arg (and the
undocumented `res` argument type), the request object can be received
prior to the draining of its data.

This is a breaking change, because it changes the signature of
onDeployment.