profiling: create profile data object at start
This is backwards compatible with the previous behaviour, but ensures
that there is a ProfileData for in-progress (started but not stopped)
cpu profiling. This lays the groundwork for the cpu profile to stop
automatically based on some condition: the client will be able to find
and poll the in-progress profiles to detect the stop, even if the
profile wasn't explicitly stopped.