allow exited processes to be resumed if matched
If the cluster outlives the mesh server then the processes could still
be alive and "resumed" by re-emitting a 'fork' message with the
correct pid/wid/pst.

This is necessary for the strong-pm docker driver to reconnect to
service instnace containers that are still running even after PM itself
has exited and restarted.