fix race condition in process update processing
If multiple updates to the same process were happening concurrently
then the last handler to call `proc.save()` would overwrite the
changes made by the other handler. This is because `proc.save()` saves
the entire object instead of only updating the changed fields.

This manifested in listenening sockets being dropped randomly, as
reported in strongloop-internal/scrum-nodeops#639, because the
listening and status:wd events are sent at almost the same time and
are sometimes processed concurrently.

To reduce the problems caused by this, `proc.updateAttributes()` is
used wherever possible so that concurrent saves to different attributes
are at least safe.

Note that there is still a problem if multiple updates to the same
attributes happen concurrently. Fortunately, that does not tend to
happen in practice.