meshctl: print a success response for set-size
Also, fixed tests so that all the assertions are counted in the plan.
Before this, there was dead test code where assertions were no longer
being hit, because ctlRequest messages are no longer being sent. The
behaviour was right, but the tests were wrong, and the tests were only
passing by accident.