test: fix test-instance-profiling
It was creating Executor models, at some time, the meshctl helper must
have started creating these executors. Anyhow, it fails because an
Executor with that ID already exists, and is unnecessary.

Also, profiles are now transmitted in the response message, so modified
test to include a 'PROFILE' string as the profile data.