fix regression in process hooks
New hooks were injected into the process update chain in f4cbe6a6 but
the default implementations did not forward their results, breaking
things.

The fix is to not rely on the hooks passing back the proc via the
callback, but do check for an error if they set one.