make InstanceActions ephemeral
Instead of persisting them to disk and waiting for that persistence to
finish before handing back the result, just use the memory connector.

Furthermore, there isn't any reason to keep them around at all, so for
the immediate future we can use a simple 10s expiration so that the
instances can GC themselves instead of just accumulating in memory
indefinitely.